아젠다 가져오기 API 작업

+++ domain/src/main/java/forutune/meeron/domain/model/Agenda.kt
@@ -4,7 +4,7 @@ import java.io.Serializable
 
 @kotlinx.serialization.Serializable
 data class Agenda(
-    val order: Long = 0,
+    val order: Int = 0,
     val name: String = "",
     val issues: List<Issue> = emptyList(),
     val fileInfos: List<FileInfo> = emptyList()
